further improvements
- remove unnecessary db clients from E2ET because script consumes data from table and print it to console
- fixed version.property path in commons
- unified way of printing supported script engines
- added missing compositekey.ben.xml demo case
- preparations maven central upload
- delete .classpath because it is ide related
- preparation demo cases for GraalVM Scripts
- removed old logging config

+++ src/main/java/com/rapiddweller/benerator/primitive/EquivalenceStringGenerator.java
@@ -77,7 +77,7 @@ public class EquivalenceStringGenerator<E> extends CardinalGenerator<E, String> 
 			getSource().reset();                            // ...and reset character selection
 			part = generateFromSource();
 		}
-		String result = createString(part.unwrap(), currentLength.intValue());
+		String result = createString(part.unwrap(), currentLength);
 		if (currentLength == 0) 
 			while (generateFromSource() != null) {
 				// for length 0, any character repetition is "",
